Fix errors after merge.
