+++ timesheet/src/main/java/com/gev/timesheet/dao/TimesheetDAO.java
@@ -1,9 +1,8 @@
-package com.dev.timesheet.dao;
+package com.gev.timesheet.dao;
 
-import com.dev.timesheet.domain.Timesheet;
+import com.gev.timesheet.domain.Timesheet;
 import org.springframework.data.mongodb.repository.MongoRepository;
 
-import java.util.List;
 import java.util.Optional;
 
 public interface TimesheetDAO extends MongoRepository<Timesheet, String> {
