+++ timesheet/src/main/java/com/gev/timesheet/dto/DTO.java
@@ -1,4 +1,4 @@
-package com.dev.timesheet.dto;
+package com.gev.timesheet.dto;
 
 public interface DTO<T> {
     void convertToDTO(T t);
